requirements.txt
Flask==0.10.1

requests
firebase_admin
line-bot-sdk
